#2712 (real 2698) : unthread VIP execution monitoring

+++ shanoir-ng-datasets/src/main/java/org/shanoir/ng/vip/output/service/OutputService.java
@@ -72,17 +72,20 @@ public class OutputService {
     public void process(ExecutionMonitoring monitoring) throws ResultHandlerException, EntityNotFoundException {
         File userImportDir = new File(this.importDir + File.separator + monitoring.getResultsLocation());
 
-        for (File archive : getArchivesToProcess(userImportDir)) {
-            File cacheFolder = new File(userImportDir.getAbsolutePath() + File.separator + FilenameUtils.getBaseName(archive.getName()));
-            List<File> outputFiles = extractTarIntoCache(archive, cacheFolder);
 
-            for (OutputHandler outputHandler : outputHandlers) {
-                if (outputHandler.canProcess(monitoring)) {
-                    LOG.info("Processing result file [{}] with [{}] output processing", archive.getAbsolutePath(), outputHandler.getClass().getSimpleName());
-                    outputHandler.manageTarGzResult(outputFiles, userImportDir, monitoring);
+        if (userImportDir.exists()) {
+            for (File archive : getArchivesToProcess(userImportDir)) {
+                File cacheFolder = new File(userImportDir.getAbsolutePath() + File.separator + FilenameUtils.getBaseName(archive.getName()));
+                List<File> outputFiles = extractTarIntoCache(archive, cacheFolder);
+
+                for (OutputHandler outputHandler : outputHandlers) {
+                    if (outputHandler.canProcess(monitoring)) {
+                        LOG.info("Processing result file [{}] with [{}] output processing", archive.getAbsolutePath(), outputHandler.getClass().getSimpleName());
+                        outputHandler.manageTarGzResult(outputFiles, userImportDir, monitoring);
+                    }
                 }
+                deleteTemporaryDirectory(cacheFolder);
             }
-            deleteTemporaryDirectory(cacheFolder);
         }
 
         // Remove processed datasets from current execution monitoring
@@ -96,6 +99,7 @@ public class OutputService {
             FileUtils.deleteDirectory(userImportDir);
         } catch (IOException e) {
             LOG.error("I/O error while deleting cache dir [{}]", userImportDir.getAbsolutePath());
+            LOG.error(e.getCause().getMessage(), e);
         }
     }
